[#96] Refactor: separate LogoutModal into file
